requirements.txt
#for frontend
streamlit
#to access gemini pro 
google-generativeai
#to load enviroment variables
python-dotenv